add code

+++ source/app/Http/Controllers/RoleController.php
@@ -20,21 +20,32 @@ class RoleController extends Controller
         $description = $request->input('desc','');
         $funtioid = $request->input('funcid','');
 
-        $role = new Role();
-        $role->name = $name;
-        $role->description = $description;
-        $role->update_date = date("Y-m-d H:i:s",time()+8*3600);
-        $role->is_deleted = 0;
-        $role->save();
-        $role_id = $role->id;
-        $arr = explode('_',$funtioid);
-        foreach($arr as $val)
+        if(!empty($name))
         {
-            $role_func = new Role_function();
-            $role_func->role_id = $role_id;
-            $role_func->function_id = $val;
-            $role_func->save();
+            $role = new Role();
+            $role->name = $name;
+            $role->description = $description;
+            $role->update_date = date("Y-m-d H:i:s",time()+8*3600);
+            $role->is_deleted = 0;
+            $role->save();
+            $role_id = $role->id;
+            $arr = explode('_',$funtioid);
+            foreach($arr as $val)
+            {
+                $role_func = new Role_function();
+                $role_func->role_id = $role_id;
+                $role_func->function_id = $val;
+                $role_func->save();
+                return array('result' => 'success');
+            }
         }
+        else
+        {
+            $data = array();
+            $data['errorcode'] = ErrorCode::AUTH.ErrorCode::INVALID_PARAM;
+            return $data;
+        }
+
     }
 
     public function updaterole($roleid,Request $request)
@@ -44,23 +55,34 @@ class RoleController extends Controller
         $funtioid = $request->input('funcid','');
 
 
-        $role_data  = Role::where('id','=',$roleid)->first();
-        $role_data->name = $name;
-        $role_data->description = $description;
-        $role_data->update_date = date("Y-m-d H:i:s",time()+8*3600);
-        $role_data->save();
-
-        $sql = "delete from role_function where role_id = $roleid";
-        DB::delete($sql);
-
-        $arr = explode('_',$funtioid);
-        foreach($arr as $val)
+        if(!empty($name))
         {
-            $role_func = new Role_function();
-            $role_func->role_id = $roleid;
-            $role_func->function_id = $val;
-            $role_func->save();
+            $role_data  = Role::where('id','=',$roleid)->first();
+            $role_data->name = $name;
+            $role_data->description = $description;
+            $role_data->update_date = date("Y-m-d H:i:s",time()+8*3600);
+            $role_data->save();
+
+            $sql = "delete from role_function where role_id = $roleid";
+            DB::delete($sql);
+
+            $arr = explode('_',$funtioid);
+            foreach($arr as $val)
+            {
+                $role_func = new Role_function();
+                $role_func->role_id = $roleid;
+                $role_func->function_id = $val;
+                $role_func->save();
+            }
+            return array('result' => 'success');
         }
+        else
+        {
+            $data = array();
+            $data['errorcode'] = ErrorCode::AUTH.ErrorCode::INVALID_PARAM;
+            return $data;
+        }
+
 
     }
 
@@ -72,12 +94,21 @@ class RoleController extends Controller
 
     public  function  deleterole($roleid)
     {
-        $sql = "update role set is_deleted = 1 where id = $roleid";
-        DB::update($sql);
-        $delete_fun = "delete from role_function where role_id = $roleid";
-        DB::delete($delete_fun);
-        $delete_user = "delete from user_role where role_id = $roleid";
-        DB::delete($delete_user);
+        if($roleid>0) {
+            $sql = "update role set is_deleted = 1 where id = $roleid";
+            DB::update($sql);
+            $delete_fun = "delete from role_function where role_id = $roleid";
+            DB::delete($delete_fun);
+            $delete_user = "delete from user_role where role_id = $roleid";
+            DB::delete($delete_user);
+            return array('result' => 'success');
+        }
+        else
+        {
+            $data = array();
+            $data['errorcode'] = ErrorCode::AUTH.ErrorCode::INVALID_PARAM;
+            return $data;
+        }
     }
 
 }
